Remove dcr/calc.py and wallet/api.py, after moving the used code elsewhere.
