+++ decred/tests/integration/dcr/test_rpc.py
@@ -203,7 +203,8 @@ cFilter414000 = ByteArray(
     "0000000e590860091d85960f114d2e457d101825e82c901465e94aa768191dcae08475a842086ad954d6"
 )
 cFilterHeader414000 = ByteArray(
-    "412f12ed5bd92df6a8b17cf396697cd44d84ada67f5ca3b1c3f23f3b7619984b"
+    # "412f12ed5bd92df6a8b17cf396697cd44d84ada67f5ca3b1c3f23f3b7619984b"
+    "a854742e7b3b8cd4f48fd2568c021e24bde9aaf07f1898cf0a9c1f133bcfe115"
 )
 blkHeader414002 = ByteArray(
     "07000000b8b8539315eb883c775cc3ef9ba0dface4dfb89def7b381a0000000000000000e"
